+++ src/DiceGameState/Bid.h
@@ -11,7 +11,7 @@
 
 @interface Bid : NSObject {
 @private
-    int playerID;
+    NSInteger playerID;
     int numberOfDice;
     int rankOfDie;
     NSArray *diceToPush;
@@ -19,13 +19,13 @@
 }
 
 @property (nonatomic, retain) NSArray *diceToPush;
-@property (nonatomic, readonly) int playerID;
+@property (nonatomic, readonly) NSInteger playerID;
 @property (nonatomic, readonly) int numberOfDice;
 @property (nonatomic, readonly) int rankOfDie;
 @property (readwrite, retain) NSString *playerName;
 
-- (id)initWithPlayerID:(int)playerIDToSet name:(NSString *)playerName dice:(int)dice rank:(int)rank;
-- (id)initWithPlayerID:(int)playerIDToSet name:(NSString *)playerName dice:(int)dice rank:(int)rank push:(NSArray *)dicePushing;
+- (id)initWithPlayerID:(NSInteger)playerIDToSet name:(NSString *)playerName dice:(int)dice rank:(int)rank;
+- (id)initWithPlayerID:(NSInteger)playerIDToSet name:(NSString *)playerName dice:(int)dice rank:(int)rank push:(NSArray *)dicePushing;
 
 - (void)dealloc;
